+++ .gitignore
@@ -7,3 +7,4 @@ yarn-error.log*
 .idea
 .npm
 tests/coverage
+sandbox/dist/
