--- demo/bitbox02-api-go.js
@@ -1,1 +1,0 @@
-../src/bitbox02-api-go.js
