--- demo/bitbox02-api.js
@@ -1,1 +1,0 @@
-../src/index.js
