--- demo/bitbox02.js
@@ -1,1 +1,0 @@
-../src/bitbox02.js
