--- demo/utils.js
@@ -1,1 +1,0 @@
-../src/utils.js
