+++ lib/bitbox02.js
@@ -6,7 +6,7 @@ Object.defineProperty(exports, "__esModule", {
 exports.getDevicePath = getDevicePath;
 exports.BitBox02API = exports.HARDENED = exports.isErrorAbort = exports.constants = void 0;
 
-require("./bitbox02-api-go.js");
+var _bitbox02ApiGo = require("./bitbox02-api-go.js");
 
 var _utils = require("./utils.js");
 
@@ -34,10 +34,10 @@ function asyncGeneratorStep(gen, resolve, reject, _next, _throw, key, arg) { try
 
 function _asyncToGenerator(fn) { return function () { var self = this, args = arguments; return new Promise(function (resolve, reject) { var gen = fn.apply(self, args); function _next(value) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, "next", value); } function _throw(err) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, "throw", err); } _next(undefined); }); }; }
 
-var api = bitbox02;
-var constants = bitbox02.constants;
+var api = _bitbox02ApiGo.bitbox02;
+var constants = _bitbox02ApiGo.bitbox02.constants;
 exports.constants = constants;
-var isErrorAbort = bitbox02.IsErrorAbort;
+var isErrorAbort = _bitbox02ApiGo.bitbox02.IsErrorAbort;
 exports.isErrorAbort = isErrorAbort;
 var HARDENED = 0x80000000;
 exports.HARDENED = HARDENED;
