+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bitbox02-api",
-  "version": "0.10.0",
+  "version": "0.11.0",
   "description": "BitBox02 JavaScript library – To enable communication from the browser to the BitBox02 hardware wallet",
   "main": "lib/index.js",
   "files": [
